Split build.gradle.kts into smaller files
