+++ skiko/buildSrc/build.gradle.kts
@@ -4,9 +4,12 @@ plugins {
 
 repositories {
     mavenCentral()
+    gradlePluginPortal()
 }
 
 dependencies {
     implementation(kotlin("stdlib"))
     compileOnly(gradleApi())
+    implementation(kotlin("gradle-plugin", "1.9.21"))
+    implementation("de.undercouch:gradle-download-task:5.4.0")
 }
